Giant refactoring of the Player API.
It all started with a rethinking of the TimingObject interface, whereby the Player now controls the TimingObject lifetime as opposed to receiving it from the caller.
This introduced the conundrum of differentiating between internally-generated TimingObject updates (such as play(), pause()) versus updates caused by the Player client.

Once that was solved(-ish, check https://github.com/chrisguttandin/timing-object/issues/229) I noticed that player velocity can be changed dynamically via a TimingObject update
but that other options such as mute / repeat need a full rebuild of the Player in order to take effect. I exposed those as dynamic attributes and updated the demo accordingly.